requirements.txt
# Core dependencies
python-nmap>=0.7.1
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
dnspython>=2.2.0
python-whois>=0.7.3
paramiko>=2.11.0
impacket>=0.10.0

# Web scanning
selenium>=4.5.0
urllib3>=1.26.0
aiohttp>=3.8.0

# Network utilities
scapy>=2.4.5
netaddr>=0.8.0
ipaddress>=1.0.23

# Reporting
jinja2>=3.1.0
markdown>=3.4.0
weasyprint>=57.0

# Utilities
colorama>=0.4.5
tqdm>=4.64.0
click>=8.1.0
pyyaml>=6.0
python-dateutil>=2.8.0

# Optional dependencies for advanced features
shodan>=1.28.0
censys>=2.1.0
virustotal-api>=1.1.11
